% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_CATCH_TYPES_COLORS}
\alias{REF_CATCH_TYPES_COLORS}
\title{The reference catch types colors}
\format{
\describe{
  \item{CATCH_TYPE_CODE}{The catch type code}
  \item{FILL}{The fill color for the identified catch type}
  \item{COLOR}{The outline color for the identified catch type}
}
}
\usage{
REF_CATCH_TYPES_COLORS
}
\description{
The reference catch types colors
}
\keyword{datasets}
