% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SPECIES_GEAR_GROUPS_COLORS}
\alias{REF_SPECIES_GEAR_GROUPS_COLORS}
\title{The reference species gear groups colors}
\format{
\describe{
  \item{SPECIES_GEAR_GROUP}{The species gear group}
  \item{FILL}{The fill color for the identified species gear group}
  \item{COLOR}{The outline color for the identified species gear group}
}
}
\usage{
REF_SPECIES_GEAR_GROUPS_COLORS
}
\description{
Includes color definitions for the 'synthetic' Others species gear group
}
\keyword{datasets}
